Change lighthouse performance score to float

+++ backend/audits/models.py
@@ -63,4 +63,4 @@ class AuditResults(BaseModel):
     wpt_metric_repeat_view_time_to_first_byte = models.IntegerField(
         blank=True, null=True
     )
-    wpt_metric_lighthouse_performance = models.IntegerField(blank=True, null=True)
+    wpt_metric_lighthouse_performance = models.FloatField(blank=True, null=True)
